+++ CHANGELOG.md
@@ -3,15 +3,18 @@
 ## 3.0
 
 ### NEXT
-* Dependency Updates:
-    * `kotlincrypto:secure-random:0.3.2` -> `kotlincrypto.random:crypto-rand:0.5.0`
-        * This fixes key generation in WASM/JS
+* PBKDF2 Support
+* HKDF Support
+* SCrypt Support
 * Add `SpecializedSymmetricEncryptionAlgorithm`
     * This allows `randomKey()` etc to operate on COSE/JWE algorithms
 * Clean up some function signatures:
     * `SymmetricKey.toJsonWebKey` now returns `KmmResult`
     * `SymmetricEncryptionAlgorithm.toJweKwAlgorithm` now returns `KmmResult`
     * `SymmetricEncryptionAlgorithm.toJweEncryptionAlgorithm` removed
+* Dependency Updates:
+    * `kotlincrypto:secure-random:0.3.2` -> `kotlincrypto.random:crypto-rand:0.5.0`
+        * This fixes key generation in WASM/JS
 
 ### 3.16.2
 * Lower Android `minSDK` to 21 (5.0 Lollipop) for all modules Except _Supreme_